requirements.txt
Django==2.2.24
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
psycopg2>=2.7,<3.0

